fifo

+++ hw/char/exynos4210_uart.c
@@ -221,7 +221,7 @@ static uint8_t fifo_retrieve(Exynos4210UartFIFO *q)
     return  ret;
 }
 
-static int fifo_elements_number(const Exynos4210UartFIFO *q)
+static size_t fifo_elements_number(const Exynos4210UartFIFO *q)
 {
     if (q->sp < q->rp) {
         return q->size - q->rp + q->sp;
@@ -230,7 +230,7 @@ static int fifo_elements_number(const Exynos4210UartFIFO *q)
     return q->sp - q->rp;
 }
 
-static int fifo_empty_elements_number(const Exynos4210UartFIFO *q)
+static size_t fifo_empty_elements_number(const Exynos4210UartFIFO *q)
 {
     return q->size - fifo_elements_number(q);
 }
